requirements.txt
asgiref==3.9.1
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==43.0.3
defusedxml==0.7.1
django-extensions==3.2.3
gunicorn==21.2.0
Django==5.2.4
djangosaml2==1.11.1
django-environ==0.12.0
django-auth-adfs==1.15.0
django-formtools==2.5.1
django-otp==1.6.1
django-phonenumber-field==8.1.0
django-two-factor-auth==1.17.0
idna==3.10
pycparser==2.22
pyopenssl==24.2.1
PyJWT==2.10.1
pypng==0.20220715.0
pysaml2==7.5.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
qrcode==7.4.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
xmlschema==2.5.1
